fix docs for kube-core
