+++ kube-core/src/metadata.rs
@@ -1,7 +1,6 @@
+//! Metadata structs used in traits, lists, and dynamic objects.
 pub use k8s_openapi::apimachinery::pkg::apis::meta::v1::{ListMeta, ObjectMeta};
 use serde::{Deserialize, Serialize};
-
-pub use crate::resource::{Resource, ResourceExt};
 
 /// Type information that is flattened into every kubernetes object
 #[derive(Deserialize, Serialize, Clone, Default, Debug, Eq, PartialEq, Hash)]
