+++ kube-core/src/resource.rs
@@ -22,7 +22,7 @@ pub trait Resource {
     /// Types that require some information at runtime should select `DynamicType`
     /// as type of this information.
     ///
-    /// See [`DynamicObject`](crate::api::DynamicObject) for a valid implementation of non-k8s-openapi resources.
+    /// See [`DynamicObject`](crate::dynamic::DynamicObject) for a valid implementation of non-k8s-openapi resources.
     type DynamicType: Send + Sync + 'static;
 
     /// Returns kind of this object
